refactor(platform): извлечь ts-platform (optimize/thumbnail) + дедуп монолита (M2 #121)
Платформо-оптимизация (optimize_for_platform_logic / generate_thumbnail_logic /
build_ffmpeg_command / validate_* / check_platform_compatibility — ffmpeg-CLI, 0 Tauri)
вынесена из монолитного video_compiler/commands/platform_optimization в крейт ts-platform →
headless, доступна агенту (шаг M2 эпика #119).

- ts-platform: business_logic + types + 18 перенесённых тестов — cargo test зелёный.
- монолит: business_logic.rs/types.rs → re-export ts_platform; Tauri-команды (commands.rs) и
  ai_metadata остаются обёртками; tests.rs удалён (живёт в крейте).

Проверено ЛОКАЛЬНО: cargo check монолита зелёный (0 ошибок); ts-platform 18 тестов.
Дальше M2: единый timeline-бинарь с субкомандой optimize поверх ts-platform.
