+++ crates/ts-platform/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "ts-platform"
+version = "0.1.0"
+edition = "2021"
+description = "Платформо-оптимизация видео Timeline Studio (optimize/thumbnail под YT/TikTok/IG/FB). Headless, ffmpeg-CLI, без Tauri. M2 #121."
+
+[lib]
+name = "ts_platform"
+path = "src/lib.rs"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+specta = { version = "2.0.0-rc.21", features = ["derive"] }
+tokio = { version = "1", features = ["process", "rt", "macros", "rt-multi-thread"] }
