+++ src-tauri/Cargo.toml
@@ -81,6 +81,7 @@ tauri-plugin-sql = { version = "2.3", features = ["sqlite"] }
 rusqlite = { version = "0.32", features = ["bundled", "vtab", "load_extension"] }
 # Video Compiler dependencies
 ts-schema = { path = "../crates/ts-schema" }
+ts-platform = { path = "../crates/ts-platform" }
 ts-analysis = { path = "../crates/ts-analysis" }
 ts-render = { path = "../crates/ts-render" }
 ts-montage = { path = "../crates/ts-montage" }
