+++ src-tauri/src/video_compiler/commands/platform_optimization/mod.rs
@@ -14,9 +14,6 @@ pub mod business_logic;
 pub mod commands;
 pub mod types;
 
-#[cfg(test)]
-mod tests;
-
 // Re-export основных типов и команд
 pub use ai_metadata_commands::*;
 pub use ai_metadata_types::*;
